Split descriptions into separate namespace files
