+++ packages/@romejs/diagnostics/descriptions/compiler.ts
@@ -1,0 +1,13 @@
+import {createDiagnosticsCategory} from "./index";
+
+// @romejs/compiler
+export const compiler = createDiagnosticsCategory({
+	CLASSES_UNSUPPORTED: {
+		category: "compile/classes",
+		message: "The classes transform doesn't know how to transform this",
+	},
+	JSX_NOT_XML: {
+		category: "compile/jsx",
+		message: "JSX is not XML",
+	},
+});
