+++ packages/@romejs/diagnostics/descriptions/pathMatch.ts
@@ -1,0 +1,6 @@
+import {createDiagnosticsCategory} from "./index";
+
+export const pathMatch = createDiagnosticsCategory({
+	INVALID_PATTERN_SEGMENT_PART: "Invalid pattern segment part",
+	INVALID_PATH_SEGMENT: "Invalid path segment",
+});
