+++ packages/@romejs/diagnostics/descriptions/v8.ts
@@ -1,0 +1,5 @@
+import {createDiagnosticsCategory} from "./index";
+
+export const v8 = createDiagnosticsCategory({
+	SYNTAX_ERROR: (message: string) => ({message, category: "v8/syntaxError"}),
+});
